User tests are passing, ride tests disabled.

+++ src/user/user.repository.ts
@@ -4,23 +4,39 @@ import { IUser } from './user.interface';
 
 export class userRepository {
   
-  static getAll(conditions?: Object, select?: string, populate?: Object[]) {
-    return User.find(conditions || {}, select || {}).populate(populate || {});
+  static getAll(conditions?: Object, select?: string, populate?: any) {
+    if (populate) {
+      return User.find(conditions || {}, select || {}).populate(populate)
+    }
+
+    return User.find(conditions || {}, select || {});
   }
 
-  static getOneByProps(conditions: Object, populate?: Object[], select?: string) {
-    return User.findOne(conditions, select || {}).populate(populate || {});
+  static getOneByProps(conditions: Object, populate?: any, select?: string) {
+    if (populate) {
+      return User.findOne(conditions, select || {}).populate(populate || {});
+    }
+
+    return User.findOne(conditions, select || {})
   }
 
   static save(user: IUser) {
     return user.save();
   }
 
-  static deleteById(id: string, populate?: Object[]) {
-    return User.findByIdAndRemove(id).populate(populate || {});
+  static deleteById(id: string, populate?: any) {
+    if (populate) {
+      return User.findByIdAndRemove(id).populate(populate || {});
+    }
+
+    return User.findByIdAndRemove(id)
   }
 
-  static updateById(id: string, update: Object, populate?: Object[]) {
-    return User.findByIdAndUpdate(id, update, { new: true }).populate(populate || {});
+  static updateById(id: string, update: any, populate?: any) {
+    if (populate) {
+      return User.findByIdAndUpdate(id, update, { new: true }).populate(populate || {});
+    }
+
+    return User.findByIdAndUpdate(id, update, { new: true });
   }
 }
